# Make DbKvs much faster for certain types of queries