@j-baker 2016-07-15
Oh, and on a row with only one version:

```

atlas=# EXPLAIN ANALYSE SELECT wrap.row_name, wrap.col_name, wrap.ts, wrap.val
        FROM gatekeeper_v2__resources wrap, (
        /* GET_LATEST_ROW_INNER (gatekeeper_v2__resources) */
             SELECT m.row_name, m.col_name, max(m.ts) as ts
             FROM gatekeeper_v2__resources m
             WHERE m.row_name =  '\x800000000000bec3'
             AND m.ts < 160000
             GROUP BY m.row_name, m.col_name ) i
        WHERE wrap.row_name = i.row_name AND wrap.col_name = i.col_name AND wrap.ts = i.ts;
                                                                                QUERY PLAN

-----------------------------------------------------------------------------------------------------------------------------------------------------------------------
---
 Nested Loop  (cost=0.83..8.92 rows=1 width=85) (actual time=0.022..0.023 rows=1 loops=1)
   ->  GroupAggregate  (cost=0.41..4.45 rows=1 width=19) (actual time=0.018..0.018 rows=1 loops=1)
         ->  Index Only Scan using pk_gatekeeper_v2__resources on gatekeeper_v2__resources m  (cost=0.41..4.43 rows=1 width=19) (actual time=0.014..0.015 rows=1 loops=
1)
               Index Cond: ((row_name = '\x800000000000bec3'::bytea) AND (ts < 160000))
               Heap Fetches: 1
   ->  Index Scan using pk_gatekeeper_v2__resources on gatekeeper_v2__resources wrap  (cost=0.41..4.45 rows=1 width=85) (actual time=0.003..0.003 rows=1 loops=1)
         Index Cond: ((row_name = m.row_name) AND (col_name = m.col_name) AND (ts = (max(m.ts))))
 Total runtime: 0.057 ms
(8 rows)

atlas=# EXPLAIN ANALYSE SELECT cells.row_name, cells.col_name, data.ts, data.val
FROM (SELECT t1.row_name AS row_name, t2.col_name AS col_name
      FROM (SELECT unnest(array[decode('800000000000bec3', 'hex')]) AS row_name) t1, (SELECT unnest(array[decode('70', 'hex')]) AS col_name) t2) cells
INNER JOIN LATERAL (SELECT ts, val FROM gatekeeper_v2__resources tbl
WHERE tbl.row_name = cells.row_name
    AND tbl.col_name = cells.col_name
    AND tbl.ts < 160000
    ORDER BY tbl.ts
    DESC LIMIT 1) data
ON true;
                                                                                      QUERY PLAN

-----------------------------------------------------------------------------------------------------------------------------------------------------------------------
---------------
 Nested Loop  (cost=0.41..44827.26 rows=10000 width=138) (actual time=0.023..0.025 rows=1 loops=1)
   ->  Result  (cost=0.00..0.51 rows=100 width=0) (actual time=0.003..0.003 rows=1 loops=1)
   ->  Nested Loop  (cost=0.41..447.26 rows=100 width=106) (actual time=0.017..0.019 rows=1 loops=1)
         ->  Result  (cost=0.00..0.51 rows=100 width=0) (actual time=0.001..0.002 rows=1 loops=1)
         ->  Limit  (cost=0.41..4.44 rows=1 width=74) (actual time=0.014..0.014 rows=1 loops=1)
               ->  Index Scan Backward using pk_gatekeeper_v2__resources on gatekeeper_v2__resources tbl  (cost=0.41..4.44 rows=1 width=74) (actual time=0.013..0.013 r
ows=1 loops=1)
                     Index Cond: ((row_name = (unnest('{"\\x800000000000bec3"}'::bytea[]))) AND (col_name = (unnest('{"\\x70"}'::bytea[]))) AND (ts < 160000))
 Total runtime: 0.048 ms
```