@j-baker 2016-07-15
to be clear, code is intentionally low quality (this is very much a POC).