@joelea 2016-08-16
Closing this as it's not going to be looked at in the immediate future, however we are explicitly tracking it in #786